+++ src/components/Form/index.ts
@@ -1,4 +1,5 @@
 export { default as AiSuggestion } from './AiSuggestion';
 export { default as FormCard } from './FormCard';
+export { default as FormErrorMessage } from './FormErrorMessage';
 export { default as MarkdownInput } from './MarkdownInput';
 export { default as PeriodInput } from './PeriodInput';
